+++ inst/templates/ad_adcm.R
@@ -4,7 +4,7 @@
 #
 # Input: cm, adsl
 library(admiral)
-library(admiral.test) # Contains example datasets from the CDISC pilot project
+library(pharmaversesdtm) # Contains example datasets from the CDISC pilot project
 library(dplyr)
 library(lubridate)
 
@@ -14,11 +14,10 @@ library(lubridate)
 # as needed and assign to the variables below.
 # For illustration purposes read in admiral test data
 
-data("admiral_cm")
+data("cm")
 data("admiral_adsl")
 
 adsl <- admiral_adsl
-cm <- admiral_cm
 
 # When SAS datasets are imported into R using haven::read_sas(), missing
 # character values from SAS appear as "" characters in R, instead of appearing