+++ inst/templates/ad_advs.R
@@ -4,7 +4,7 @@
 #
 # Input: adsl, vs
 library(admiral)
-library(admiral.test) # Contains example datasets from the CDISC pilot project
+library(pharmaversesdtm) # Contains example datasets from the CDISC pilot project
 library(dplyr)
 library(lubridate)
 library(stringr)
@@ -15,11 +15,10 @@ library(stringr)
 # as needed and assign to the variables below.
 # For illustration purposes read in admiral test data
 
-data("admiral_vs")
+data("vs")
 data("admiral_adsl")
 
 adsl <- admiral_adsl
-vs <- admiral_vs
 
 # When SAS datasets are imported into R using haven::read_sas(), missing
 # character values from SAS appear as "" characters in R, instead of appearing